# public/docs/api.yaml
openapi: 3.0.0
info:
  title: HRM PRO
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /education:
    post:
      tags:
        - hr > education
      summary: create single education
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 1
                degree: Bachelor's
                institution: Example University
                fieldOfStudy: Computer Science
                result: 3.5 GPA
                studyStartDate: 2018-09-01
      responses:
        '201':
          description: Successful response
          content:
            application/json: {} 
        '500':
          description: server side error
          content:
            application/json: {}
    get:
      tags:
        - hr > education
      summary: get all education by all or by paginated and by status query 
      parameters:
        - name: query
          in: query
          schema:
            type: string
            enum: ['all']
            example: all
        - name: status
          in: query
          schema:
            type: string
            enum: ['true', 'false']
        - name: page
          in: query
          schema:
            type: integer
        - name: count
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '500':
          description: server side error
          content:
            application/json: {}
...
  /email:
    get:
      tags:
        - Email
      summary: get email
      parameters:
        - name: type
          in: query
          schema:
            type: string
          example: saleinvoice
        - name: id
          in: query
          schema:
            type: integer
          example: '1'
        - name: emailaddress
          in: query
          schema:
            type: string
          example: rshaon09@gmail.com
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '500':
          description: Server Side Error
          content:
            application/json: {}
        '400':
          description: config error
          content:
            application/json: {}
...